+++ packages/components/src/components/sort-handle/sort-handle.tsx
@@ -103,8 +103,6 @@ export class SortHandle extends LitElement {
 
   /**
    * Determines where the component will be positioned relative to the container element.
-   *
-   * @default "bottom-start"
    */
   @property({ reflect: true }) placement: MenuPlacement = defaultMenuPlacement;