+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.0-next.82](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.81...@esri/calcite-components-react@5.0.0-next.82) (2026-01-27)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.81](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.80...@esri/calcite-components-react@5.0.0-next.81) (2026-01-27)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.80](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.79...@esri/calcite-components-react@5.0.0-next.80) (2026-01-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.79](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.78...@esri/calcite-components-react@5.0.0-next.79) (2026-01-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.78](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.77...@esri/calcite-components-react@5.0.0-next.78) (2026-01-22)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.0.0-next.77](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.76...@esri/calcite-components-react@5.0.0-next.77) (2026-01-22)
 
 **Note:** Version bump only for package @esri/calcite-components-react